Split Name::Format tests out of name_test.rb (#4708)
Moves the 9 tests exercising Name::Format (display_name/display_name_brief_authors/
display_name_without_authors, unknown?/known?, make_sure_names_are_bolded_correctly,
sensu_stricto, imageless?, brief_author) into test/models/name/format_test.rb,
matching app/models/name/format.rb.

test_format_autonym stayed with the formatting tests in the old file's layout,
but Name.format_autonym is defined in app/models/name/parse.rb, not format.rb -
moved it into Name::ParseTest instead (missed in the earlier Parse split since
it lived outside that split's section boundaries).

require_relative'd from name_test.rb so running that file alone still runs
the full original suite.
